Refactored timestamps to dedicated db collection
