+++ gens/db/__init__.py
@@ -1,3 +1,2 @@
-from .db import RecordType, VariantCategory
 from .db import init_database_connection as init_database
-from .db import query_records_in_region, query_variants
+from .annotation import query_records_in_region, query_variants, RecordType, VariantCategory, register_data_update, get_timestamps
